restructure powchain (#1983)
